Closes #60
Move request/responses classes to their own classes
