+++ backend/src/contracts/Responses/ListResponses.cs
@@ -1,0 +1,4 @@
+using core;
+
+namespace contracts.Responses;
+public record GetListResponse(int Id, string Name, IEnumerable<TodoItem> Items);
